:ambulance: Atualizacao de deletar imagem

+++ app/Models/Corporate.php
@@ -47,9 +47,11 @@ class Corporate extends Init
         parent::boot();
 
         static::deleting(function (Model $model) {
-            $file = 'images/' . $model->image->uri ?? '';
-            if (Storage::disk('public')->exists($file)) {
-                Storage::disk('public')->delete($file);
+            if (!empty($model->image?->uri)) {
+                $file = 'images/' . $model->image->uri ?? '';
+                if (Storage::disk('public')->exists($file)) {
+                    Storage::disk('public')->delete($file);
+                }
             }
             $model->image()->delete();
         });
